requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   groq
    #   httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
deprecation==2.1.0
    # via lancedb
distro==1.9.0
    # via
    #   google-genai
    #   groq
duckdb==1.4.3
    # via ai-data-analyst-agent (pyproject.toml)
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via langchain-google-genai
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.29.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.47.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-genai==1.57.0
    # via langchain-google-genai
google-generativeai==0.8.6
    # via ai-data-analyst-agent (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
groq==0.37.1
    # via langchain-groq
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   groq
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
lance-namespace==0.4.5
    # via lancedb
lance-namespace-urllib3-client==0.4.5
    # via lance-namespace
lancedb==0.26.1
    # via ai-data-analyst-agent (pyproject.toml)
langchain-core==1.2.7
    # via
    #   langchain-google-genai
    #   langchain-groq
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==4.1.3
    # via ai-data-analyst-agent (pyproject.toml)
langchain-groq==1.1.1
    # via ai-data-analyst-agent (pyproject.toml)
langgraph==1.0.6
    # via ai-data-analyst-agent (pyproject.toml)
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.6
    # via langgraph
langgraph-sdk==0.3.2
    # via langgraph
langsmith==0.6.2
    # via langchain-core
markupsafe==3.0.3
    # via jinja2
narwhals==2.15.0
    # via
    #   altair
    #   plotly
numpy==2.4.1
    # via
    #   lancedb
    #   pandas
    #   pydeck
    #   streamlit
openpyxl==3.1.5
    # via ai-data-analyst-agent (pyproject.toml)
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
overrides==7.7.0
    # via lancedb
packaging==25.0
    # via
    #   altair
    #   deprecation
    #   lancedb
    #   langchain-core
    #   langsmith
    #   plotly
    #   streamlit
pandas==2.3.3
    # via
    #   ai-data-analyst-agent (pyproject.toml)
    #   streamlit
pillow==12.1.0
    # via streamlit
plotly==6.5.1
    # via ai-data-analyst-agent (pyproject.toml)
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via
    #   lancedb
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   ai-data-analyst-agent (pyproject.toml)
    #   google-genai
    #   google-generativeai
    #   groq
    #   lance-namespace-urllib3-client
    #   lancedb
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyparsing==3.3.1
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   lance-namespace-urllib3-client
    #   pandas
python-dotenv==1.2.1
    # via ai-data-analyst-agent (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-genai
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   google-genai
    #   groq
sqlglot==28.5.0
    # via ai-data-analyst-agent (pyproject.toml)
streamlit==1.52.2
    # via ai-data-analyst-agent (pyproject.toml)
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via
    #   google-generativeai
    #   lancedb
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   google-genai
    #   google-generativeai
    #   groq
    #   grpcio
    #   lance-namespace-urllib3-client
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via
    #   lance-namespace-urllib3-client
    #   requests
uuid-utils==0.13.0
    # via
    #   langchain-core
    #   langsmith
watchdog==6.0.0
    # via streamlit
websockets==15.0.1
    # via google-genai
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
